pyproject.toml
[project]
name = "depvis-nuget"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",

]

[project.scripts]
depvis = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"